Imported r195: Removed PrintInChunks for ancient VMS bug

+++ src/libelectron/factmngr.c
@@ -1482,7 +1482,7 @@ globle void *GetNextFactInScope(
 globle void EnvGetFactPPForm(
   void *theEnv,
   char *buffer,
-  unsigned bufferLength,
+  size_t bufferLength,
   void *theFact)
   {
    OpenStringDestination(theEnv,"FactPPForm",buffer,bufferLength);
